+++ src/app/(list)/[teamId]/task-lists/[listId]/task-detail/[id]/_components/comments/comment-input.tsx
@@ -12,7 +12,7 @@ const CommentInput = () => {
   };
 
   return (
-    <div className="relative border-y-[0.2px] border-text-disabled">
+    <form className="relative border-y-[0.2px] border-text-disabled">
       <input
         type="text"
         placeholder="댓글을 달아주세요"
@@ -27,7 +27,7 @@ const CommentInput = () => {
         className="absolute right-3 top-1/2 -translate-y-1/2 text-gray-400 hover:text-white"
         disabled={!inputValue}
       />
-    </div>
+    </form>
   );
 };
